// store/client/src/components/PopUps/Add/AddInfo/EField.ts
enum EField {
  reset = 'reset',
  infoImage = 'infoImage',
  infoImageUrl = 'infoImageUrl',
  infoTitle = 'infoTitle',
  infoHeader = 'infoHeader',
  infoListTitle = 'infoListTitle',
  infoListItems = 'infoListItems',
  infoParagraphs = 'infoParagraphs',
  fetch = 'fetch',
}

export default EField;